do marashalling in goroutine, outside of the fast path locked loop

+++ exporter/chqk8sentitygraphexporter/internal/objecthandler/autoscaling/v2/hpa.go
@@ -40,7 +40,7 @@ func ConvertHPA(config *converterconfig.Config, us unstructured.Unstructured) (b
 
 	minReplicas := int32(1)
 	if k8sobj.Spec.MinReplicas != nil {
-		minReplicas = int32(*k8sobj.Spec.MinReplicas)
+		minReplicas = *k8sobj.Spec.MinReplicas
 	}
 	hpa := &graphpb.AutoscalingHpaSummary{
 		BaseObject: baseobj.Make(config, &us, us.GetAPIVersion(), us.GetKind()),
